+++ src/app.module.ts
@@ -11,8 +11,8 @@ import {
   NotificationService,
   AlkemioClientAdapterModule,
 } from '@src/services';
-import { AlkemioClientModule, NotifmeModule } from '@src/wrappers';
-import { NotificationRecipientsAdapterModule } from './services/notification-recipients-adapter/notification.recipients.adapter.module';
+import { AlkemioClientModule, NotifmeModule } from '@src/services/external';
+import { NotificationRecipientsAdapterModule } from './services/application/notification-recipients-adapter/notification.recipients.adapter.module';
 
 @Module({
   imports: [
