+++ src/services/application/alkemio-client-adapter/alkemio.client.adapter.spec.ts
