+++ src/services/application/template-to-credential-mapper/utils/utils.spec.ts
