+++ src/services/domain/notification/notification.service.ts
@@ -2,7 +2,7 @@ import { Injectable, Inject, LoggerService } from '@nestjs/common';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import NotifmeSdk, { NotificationStatus } from 'notifme-sdk';
 import { LogContext, NOTIFICATIONS_PROVIDER } from '@src/common';
-import { ApplicationNotificationBuilder } from '../../event-handlers/application-notification-builder/application.notification.builder';
+import { ApplicationNotificationBuilder } from '../application-notification-builder/application.notification.builder';
 import { ApplicationCreatedEventPayload } from '@src/types/application.created.event.payload';
 
 @Injectable()
