+++ src/services/index.ts
@@ -1,4 +1,4 @@
-export * from './alkemio-client-adapter';
-export * from '../event-handlers/application-notification-builder/application.notification.builder';
-export * from './notification/notification.service';
-export * from './notification-recipients-adapter/';
+export * from './application/alkemio-client-adapter';
+export * from './domain/application-notification-builder/application.notification.builder';
+export * from './domain/notification/notification.service';
+export * from './application/notification-recipients-adapter';
