+++ src/types/index.ts
@@ -1,1 +1,2 @@
 export * from './user.registration.event.payload';
+export * from './application.created.event.payload';
